requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.4
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-celery-results==2.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-formtools==2.2
django-import-export==2.4.0
django-rest-framework==0.1.0
django-reversion==3.0.9
django-utils-six==2.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
future==0.18.2
httplib2==0.9.2
importlib-metadata==4.5.0
install==1.3.4
kombu==5.1.0
librabbitmq==2.0.0
MarkupPy==1.14
msgpack==1.0.2
odfpy==1.4.1
openpyxl==3.0.7
prompt-toolkit==3.0.19
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
reversion==0.2
six==1.10.0
sqlparse==0.4.1
tablib==3.0.0
tfos==2.3.3
toml==0.10.2
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
xadmin-x==0.7.4
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
